requirements.txt
mysql-connector-python>=8.0.0
paramiko>=3.0.0